Moving NemaGenerator to separate file / reducing shaping to Y direction only
