+++ Core/PrimaryGenerator.cpp
@@ -30,155 +30,6 @@
 #include <Randomize.hh>
 #include <globals.hh>
 
-NemaPoint::NemaPoint()
-{
-  G4ThreeVector temp(0,0,0);
-  position = temp;
-  lifetime = MaterialParameters::fTauBulk;
-  is3GAllowed = false;
-  G4ThreeVector temp2(0.1,0.1,0.1);
-  sizeOfPoint = temp2;
-  orientationOfPoint = temp;
-  shapeOfPointInX = temp;
-  shapeOfPointInY = temp;
-}
-
-void NemaGenerator::AddPoint(int pointID)
-{
-  auto search = fIDPointsConnection.find(pointID);
-  if (search == fIDPointsConnection.end()) {
-    int genPointSize = fGeneratedPoints.size();
-    NemaPoint newPoint;
-    fGeneratedPoints.push_back(newPoint);
-    fIDPointsConnection.insert(std::make_pair(pointID, genPointSize));
-  }
-}
-
-void NemaGenerator::AddPointWeight(int pointID, int weight)
-{
-  unsigned removedElements = 0;
-  if (weight == 0) {
-    for (unsigned i=0; i<fWeightPositions.size(); i++) {
-      if (fWeightPositions.at(i-removedElements) == pointID) {
-        fWeightPositions.erase(fWeightPositions.begin()+i-removedElements);
-        removedElements++;
-      }
-    }
-  } else {
-    int count = 0;
-    for (unsigned i=0; i<fWeightPositions.size(); i++) {
-      if (fWeightPositions.at(i) == pointID)
-        count++;
-    }
-    if (weight >= count) {
-      for (int i=0; i<weight-count; i++)
-        fWeightPositions.push_back(pointID);
-    } else {
-      int nmbOfPointToErase = count - weight;
-      for (unsigned i=0; i<fWeightPositions.size(); i++) {
-        if (fWeightPositions.at(i-removedElements) == pointID) {
-          if ((int)removedElements < nmbOfPointToErase) {
-            fWeightPositions.erase(fWeightPositions.begin()+i-removedElements);
-            removedElements++;
-          }
-        }
-      }
-    }
-  }
-}
-
-void NemaGenerator::SetOnePointOnly(int pointID)
-{
-  NemaPoint tempPoint = GetPoint(pointID);
-  Clear();
-  fGeneratedPoints.push_back(tempPoint);
-  fWeightPositions.push_back(pointID);
-  fIDPointsConnection.insert(std::make_pair(pointID, 0));
-}
-
-bool NemaGenerator::DoesPointExistAlready(int pointID) 
-{
-  auto search = fIDPointsConnection.find(pointID);
-  if (search == fIDPointsConnection.end()) {
-    return false;
-  } else {
-    return true;
-  }
-}
-
-NemaPoint NemaGenerator::GetPoint(int pointID) const
-{
-  auto search = fIDPointsConnection.find(pointID);
-  if (search != fIDPointsConnection.end()) {
-    return fGeneratedPoints.at(fIDPointsConnection.at(pointID));
-  } else {
-    NemaPoint newPoint;
-    return newPoint;
-  }
-}
-
-NemaPoint NemaGenerator::GetRandomPoint() const
-{
-  if (fGeneratedPoints.size() == 1)
-    return fGeneratedPoints.at(0);
-  
-  unsigned indexOfPoint = rand() % (int)fWeightPositions.size();
-  int pointID = fWeightPositions.at(indexOfPoint);
-  return GetPoint(pointID);
-}
-
-void NemaGenerator::GenerateElipseXNorm(int pointID)
-{
-  Int_t pointNumber = 1000;
-  Double_t thetaStep = M_PI/(2*pointNumber);
-  Double_t maxVal = 1;
-  unsigned pointVectorID = fIDPointsConnection.at(pointID);
-  double xMax = fGeneratedPoints.at(pointVectorID).shapeOfPointInX.getX();
-  double zMax = fGeneratedPoints.at(pointVectorID).sizeOfPoint.getZ();
-  if (fabs(xMax) > fabs(zMax)) {
-    maxVal = fabs(xMax);
-  }
-  else {
-    maxVal = fabs(zMax);
-  }
-  Double_t arg = 0, val = 0;
-  TGraph* newGraph = new TGraph(pointNumber);
-  for (Int_t i=0; i<pointNumber; i++) {
-    arg = thetaStep*i;
-    val = sqrt(pow(xMax*sin(thetaStep*i),2) + pow(zMax*cos(thetaStep*i),2))/maxVal;
-    newGraph -> SetPoint(i, arg, val);
-  }
-  fGeneratedPoints.at(pointVectorID).elipseXNorm = new TF1("elipseY", "pol8", 0, M_PI/2);
-  newGraph -> Fit(fGeneratedPoints.at(pointVectorID).elipseXNorm, "Q");
-  delete newGraph;
-}
-
-void NemaGenerator::GenerateElipseYNorm(int pointID)
-{
-  Int_t pointNumber = 1000;
-  Double_t thetaStep = M_PI/(2*pointNumber);
-  Double_t maxVal = 1;
-  unsigned pointVectorID = fIDPointsConnection.at(pointID);
-  double yMax = fGeneratedPoints.at(pointVectorID).shapeOfPointInY.getX();
-  double zMax = fGeneratedPoints.at(pointVectorID).sizeOfPoint.getZ();
-  if (fabs(yMax) > fabs(zMax)) {
-    maxVal = fabs(yMax);
-  }
-  else {
-    maxVal = fabs(zMax);
-  }
-  Double_t arg = 0, val = 0;
-  TGraph* newGraph = new TGraph(pointNumber);
-  for (Int_t i=0; i<pointNumber; i++) {
-    arg = thetaStep*i;
-    val = sqrt(pow(yMax*sin(thetaStep*i),2) + pow(zMax*cos(thetaStep*i),2))/maxVal;
-    newGraph -> SetPoint(i, arg, val);
-  }
-  fGeneratedPoints.at(pointVectorID).elipseYNorm = new TF1("elipseY", "pol8", 0, M_PI/2);
-  newGraph -> Fit(fGeneratedPoints.at(pointVectorID).elipseYNorm, "Q");
-  delete newGraph;
-}
-
 PrimaryGenerator::PrimaryGenerator() : G4VPrimaryGenerator() {}
 
 PrimaryGenerator::~PrimaryGenerator() {}
@@ -583,70 +434,8 @@ void PrimaryGenerator::GenerateNema(G4Event* event, NemaGenerator* nemaGen)
   
   G4ThreeVector nemaPosition = nemaPoint.position;
   G4ThreeVector vtxPosition = VertexUniformInCylinder(nemaPoint.sizeOfPoint.getX(), nemaPoint.sizeOfPoint.getZ());
-  if (nemaPoint.shapeOfPointInX.getX() != 0) {
-    double previousX = vtxPosition.getX();
-    double directionX = nemaPoint.shapeOfPointInX.getX();
-    double cutOffX = 1-nemaPoint.shapeOfPointInX.getZ();
-    bool badElipsePoint = true;
-    while (badElipsePoint) {
-      double randNum = G4UniformRand();
-      double randNum2 = cutOffX*(M_PI * G4UniformRand() - M_PI/2);
-      if (randNum <= nemaPoint.elipseYNorm -> Eval(fabs(randNum2))) {
-        G4double xP = directionX*cos(randNum2);
-        G4double zP = nemaPoint.sizeOfPoint.getZ()*sin(randNum2);
-        double signZ = zP/fabs(zP);
-        double theta = acos(xP/fabs(directionX));
-        xP = fabs(directionX)*cos(theta);
-        zP = signZ*nemaPoint.sizeOfPoint.getZ()*sin(theta);
-        xP = xP + previousX*cos(theta);
-        zP = zP + signZ*previousX*sin(theta);
-        vtxPosition.setX(xP);
-        vtxPosition.setZ(zP);
-        badElipsePoint = false;
-      }
-    }
-  }
-  if (nemaPoint.shapeOfPointInY.getX() != 0) {
-    double previousY = vtxPosition.getY();
-    double directionY = nemaPoint.shapeOfPointInY.getX();
-    double cutOffY = 1-nemaPoint.shapeOfPointInY.getZ();
-    bool badElipsePoint = true;
-    while (badElipsePoint) {
-      double randNum = G4UniformRand();
-      double randNum2 = cutOffY*(M_PI * G4UniformRand() - M_PI/2);
-      if (randNum <= nemaPoint.elipseYNorm -> Eval(fabs(randNum2))) {
-        G4double yP = directionY*cos(randNum2);
-        G4double zP = nemaPoint.sizeOfPoint.getZ()*sin(randNum2);
-        double signZ = zP/fabs(zP);
-        double theta = acos(yP/fabs(directionY));
-        yP = fabs(directionY)*cos(theta);
-        zP = signZ*nemaPoint.sizeOfPoint.getZ()*sin(theta);
-        yP = yP + previousY*cos(theta);
-        zP = zP + signZ*previousY*sin(theta);
-        vtxPosition.setY(yP);
-        vtxPosition.setZ(zP);
-        badElipsePoint = false;
-      }
-    }
-  }
-  if (nemaPoint.orientationOfPoint.mag() > 0) {
-    double theta = nemaPoint.orientationOfPoint.getX() * M_PI/180.;
-    double phi = nemaPoint.orientationOfPoint.getY() * M_PI/180.;
-    if (theta != 0. || phi != 0.) {
-      TVector3 normal(cos(theta)*sin(phi), sin(theta)*sin(phi), cos(phi));
-      normal = normal.Unit();
-      TVector3 z(0.,0.,1.);
-      TVector3 rotAxis = z.Cross(normal);
-      double angle = normal.Angle(z); // radians
-      TRotation rot;
-      rot.Rotate(angle, rotAxis);
-      TVector3 tempVector(vtxPosition.getX(), vtxPosition.getY(), vtxPosition.getZ());
-      tempVector = rot*tempVector;
-      vtxPosition.setX(tempVector.X());
-      vtxPosition.setY(tempVector.Y());
-      vtxPosition.setZ(tempVector.Z());
-    }
-  }
+  vtxPosition = nemaGen->GetPointShapedInY(vtxPosition, nemaPoint);
+  vtxPosition = nemaGen->GetRotatedPoint(vtxPosition, nemaPoint);
   vtxPosition = vtxPosition + nemaPosition;
 
   double lifetime = nemaPoint.lifetime;
