+++ Info/PrimaryGeneratorActionMessenger.h
@@ -62,7 +62,6 @@ private:
   G4UIcmdWithAnInteger* fNemaSetPosition3GOption = nullptr;
   G4UIcmdWithAString* fNemaSetPositionCylinderSize = nullptr;
   G4UIcmdWithAString* fNemaSetPositionCylinderRotation = nullptr;
-  G4UIcmdWithAString* fNemaSetPositionCylinderShapeX = nullptr;
   G4UIcmdWithAString* fNemaSetPositionCylinderShapeY = nullptr;
   G4UIcmdWith3VectorAndUnit* fSetChamberCenter = nullptr;
   G4UIcmdWithADoubleAndUnit* fSetChamberEffectivePositronRadius = nullptr;
